Day 04
I solved the first part earlier this afternoon. I decided to leverage
the constant lookup time of a Set to quickly determine if the current
row, the one above, or the one below contain a neighbor. Instead of
tracking a grid, I used a two dimensional array to track the indices at
which rolls are present.

I came back to the second part later in the evening. I did a little
refactoring for code reuse since I could make use of most of my solution
for part one. I essentially made it so that I could check for what could
be removed, remove them, then repeat. I repeat until there is nothing
left to remove. The efficiency could be better I think. I believe my
solution was in the 3-4s range.